# algoAPI/swagger.yaml
openapi: 3.0.3
info:
  title: Algo API
  description: >-
    Documentation for Algo API. REST API service, offering algorithm execution
    on request. Additionaly, input data can be saved to the service and used
    again. Most paths require a user account, which is managed through the API
    aswell.
  version: '1.0'
  x-logo:
    url: ''
servers:
  - url: 'https://algo-algo.herokuapp.com/'
...
paths:
  '/request/{algorithm}':
    post:
      tags:
        - request
      summary: Returns the result of requested algorithm
      description: >-
        Used for making algorithm requests. Can be called with or without
        authentication. If called with authentication, result is assigned to the
        user.
      operationId: requestAlgorithm
      parameters:
        - name: algorithm
          in: path
          description: Name of the requested algorithm
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: quickSort
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error parsing input
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - {}
        - JWTAuth: []
  '/request/{algorithm}/{data_id}':
    post:
      tags:
        - request
      summary: Returns the result of requested algorithm with saved data
      description: >-
        Used for making algorithm requests with input data, saved in the
        database. Parameters must be provided in the body, if not, an empty array is used.
      operationId: requestAlgorithmWithDataByID
      parameters:
        - name: algorithm
          in: path
          description: Name of the requested algorithm
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: quickSort
        - name: data_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      requestBody:
        content:
          application/json:
              schema:
                type: object
                properties:
                  parameters:
                    type: array
                    items:
                      type: string
                example:
                  parameters:
                    - descending
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error parsing input
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid or expired jwt
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
  '/request/{request_id}':
    get:
      tags:
        - request
      summary: Returns the request
      description: >-
        Used for fetching a request from the database. Can be used with or
        without authorization. Only unassigned requests (user id
        000000000000000000000000) can be requested without authorization.
      operationId: getRequestByID
      parameters:
        - name: request_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Not authorized to access request
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - {}
        - JWTAuth: []
    delete:
      tags:
        - request
      summary: Deletes the request
      operationId: deleteRequestByID
      description: Used for deleting a request from the database.
      parameters:
        - name: request_id
          in: path
          required: true
          deprecated: false
          style: simple
          explode: false
          schema:
            type: string
          example: 5f3348e14d9b386e025681a1
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid or expired jwt
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
      security:
        - JWTAuth: []
...
  /user:
    post:
      tags:
        - user
      summary: Creates a user
      description: 'Used for creating a user account, returns a JWT token.'
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              username: test
              email: test@test.com
              password: test
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/registration_response'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Invalid id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Error running algorithm
...
components:
  schemas:
    Request:
      type: object
      properties:
        id:
          type: string
        algorithm:
          type: string
        input:
          type: string
        parameters:
          type: array
          items:
            type: string
        output:
          type: string
        requested:
          type: string
        completed:
          type: string
        executionTime:
          type: string
        user:
          type: string
      example:
        id: 5fe9817753590b33f7a82d62
        algorithm: SelectionSort
        input: '3,0,33,11,2202,-200'
        parameters:
          - descending
        requested: '2020-12-28T07:58:16+01:00'
        completed: '2020-12-28T07:58:17+01:00'
        executionTime: 0s
        user: '000000000000000000000000'
    Error:
      required:
        - message
      type: object
      properties:
        message:
          type: string
          description: Error message indicating the issue
      example:
        message: User not found
    User:
      required:
        - email
        - password
        - username
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        username: test
        email: test@test.com
        password: test
...
    body:
      required:
        - data
      properties:
        data:
          type: string
        parameters:
          type: array
          items:
            type: string
      example:
        input: '3,0,33,11,2202,-200'
        parameters:
          - descending
    registration_response:
      required:
        - created
        - email
        - id
        - username
      properties:
        userID:
          type: string
        token:
          type: string
        username:
          type: string
      example:
        userID: 5fe98a1453590b33f7a82d64
        token: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MDkxNDQzNTYsInVzZXJfaWQiOiI1ZmU5OGExNDUzNTkwYjMzZjdhODJkNjQiLCJ1c2VyX3VzZXJuYW1lIjoidGVzdCJ9.jgLcRffdVVsOzdfGDl09pcQXruEnpolNuyatPp8CflM
        username: test
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT